Refactor protocol func names
