+++ PokemonTcg/Data/Local/LocalPokemonDataSource.swift
@@ -1,5 +1,5 @@
 //
-//  LocalPokemonRepository.swift
+//  LocalPokemonRepositoryImpl.swift
 //  PokemonTcg
 //
 //  Created by Cláudio Costa on 01/07/2025.
@@ -7,8 +7,22 @@
 
 import Foundation
 
-protocol LocalPokemonDataSource: Sendable {
+protocol LocalPokemonDataSourceProtocol: Sendable {
     func getCachedPokemon(for filter: PokemonFilter) async -> [PokemonCard]?
     func save(pokemon: [PokemonCard], for filter: PokemonFilter) async
 }
 
+actor LocalPokemonDataSource: LocalPokemonDataSourceProtocol {
+    private var cache: [PokemonFilter: [PokemonCard]] = [:]
+
+    func getCachedPokemon(for filter: PokemonFilter) async -> [PokemonCard]? {
+        return cache[filter]
+    }
+
+    func save(pokemon: [PokemonCard], for filter: PokemonFilter) async {
+        guard !pokemon.isEmpty else {
+            return
+        }
+        cache[filter] = pokemon
+    }
+}
