--- PokemonTcg/Domain/Repository/PokemonRepository.swift
@@ -1,10 +1,0 @@
-//
-//  File.swift
-//  PokemonTcg
-//
-//  Created by Cláudio Costa on 30/06/2025.
-//
-
-protocol PokemonRepository: Sendable {
-    func fetchPokemons(filter: PokemonFilter) async throws -> [PokemonCard]
-}
