+++ PokemonTcg/PokemonContainer.swift
@@ -11,17 +11,17 @@ import Foundation
 final class PokemonContainer {
     static let shared = PokemonContainer()
 
-    let localDataSource: LocalPokemonDataSource
-    let remoteDataSource: RemotePokemonDataSource
-    let repository: PokemonRepository
+    let localDataSource: LocalPokemonDataSourceProtocol
+    let remoteDataSource: RemotePokemonDataSourceProtocol
+    let repository: PokemonRepositoryProtocol
     let fetchPokemonCardsUseCase: FetchPokemonCardsUseCase
     let networkManager: PokemonNetworkManager
 
     private init() {
         self.networkManager = PokemonNetworkManager()
-        self.localDataSource = LocalPokemonDataSourceImpl()
-        self.remoteDataSource = RemotePokemonDataSourceImpl(networkManager: networkManager)
-        self.repository = PokemonRepositoryImpl(local: localDataSource, remote: remoteDataSource)
+        self.localDataSource = LocalPokemonDataSource()
+        self.remoteDataSource = RemotePokemonDataSource(networkManager: networkManager)
+        self.repository = PokemonRepository(local: localDataSource, remote: remoteDataSource)
         self.fetchPokemonCardsUseCase = FetchPokemonCardsUseCase(repository: repository)
     }
 }
